+++ src/sentry/native/native_sdk.cpp
@@ -1,10 +1,10 @@
 #include "native_sdk.h"
 
-#include "godot_cpp/core/error_macros.hpp"
 #include "sentry.h"
 #include "sentry/contexts.h"
 #include "sentry/environment.h"
 #include "sentry/level.h"
+#include "sentry/native/native_event.h"
 #include "sentry/native/native_util.h"
 #include "sentry_options.h"
 
@@ -66,23 +66,6 @@ inline String _uuid_as_string(sentry_uuid_t p_uuid) {
 	char str[37];
 	sentry_uuid_as_string(&p_uuid, str);
 	return str;
-}
-
-sentry_level_t _level_as_native(sentry::Level p_level) {
-	switch (p_level) {
-		case sentry::Level::LEVEL_DEBUG:
-			return SENTRY_LEVEL_DEBUG;
-		case sentry::Level::LEVEL_INFO:
-			return SENTRY_LEVEL_INFO;
-		case sentry::Level::LEVEL_WARNING:
-			return SENTRY_LEVEL_WARNING;
-		case sentry::Level::LEVEL_ERROR:
-			return SENTRY_LEVEL_ERROR;
-		case sentry::Level::LEVEL_FATAL:
-			return SENTRY_LEVEL_FATAL;
-		default:
-			ERR_FAIL_V_MSG(SENTRY_LEVEL_ERROR, "SentrySDK: Internal error - unexpected level value. Please open an issue.");
-	}
 }
 
 } // unnamed namespace
@@ -148,7 +131,7 @@ void NativeSDK::add_breadcrumb(const String &p_message, const String &p_category
 
 String NativeSDK::capture_message(const String &p_message, Level p_level, const String &p_logger) {
 	sentry_value_t event = sentry_value_new_message_event(
-			_level_as_native(p_level),
+			native::level_to_native(p_level),
 			p_logger.utf8().get_data(),
 			p_message.utf8().get_data());
 	last_uuid = sentry_capture_event(event);
@@ -184,6 +167,23 @@ String NativeSDK::capture_error(const String &p_type, const String &p_value, Lev
 	sentry_value_set_by_key(stack_trace, "frames", frames);
 	sentry_value_set_by_key(exception, "stacktrace", stack_trace);
 	sentry_event_add_exception(event, exception);
+	last_uuid = sentry_capture_event(event);
+	return _uuid_as_string(last_uuid);
+}
+
+Ref<SentryEvent> NativeSDK::create_event() {
+	sentry_value_t event_value = sentry_value_new_event();
+	Ref<SentryEvent> event = memnew(NativeEvent(event_value));
+	return event;
+}
+
+String NativeSDK::capture_event(const Ref<SentryEvent> &p_event) {
+	last_uuid = sentry_uuid_nil();
+	ERR_FAIL_COND_V_MSG(p_event.is_null(), _uuid_as_string(last_uuid), "Sentry: Can't capture event - event object is null.");
+	NativeEvent *native_event = Object::cast_to<NativeEvent>(p_event.ptr());
+	ERR_FAIL_NULL_V(native_event, _uuid_as_string(last_uuid)); // Sanity check - this should never happen.
+	sentry_value_t event = native_event->get_native_value();
+	sentry_value_incref(event); // Keep ownership.
 	last_uuid = sentry_capture_event(event);
 	return _uuid_as_string(last_uuid);
 }